Extracts EscapeHatchModel+Previews
